% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_compute.R
\name{read.lists_from_files}
\alias{read.lists_from_files}
\title{Read list of elements from files in lists.}
\usage{
read.lists_from_files(list_paths, list_names)
}
\arguments{
\item{list_paths}{A list of strings representing paths of files.}

\item{list_names}{A list of strings as names of sets.}
}
\value{
A list of list of character elements from files in \code{list_paths}.
Each list contains the elementos of the files.
}
\description{
Read list of elements from files in lists.
}
\examples{
\dontrun{
read.lists_from_files(c('genes_1.txt', '/home/user/Documents/bio/genes_2.txt'), c("g1", "g2"))
read.lists_from_files(c('genes_1.txt', 'D:\\\\bio\\\\genes_2.txt', 'genes_3.txt'), c("g1", "g2", "g3"))
}
}
