% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_compute.R
\name{venn.compute_plot}
\alias{venn.compute_plot}
\title{Compute specific elements of Venn diagram and plot.}
\usage{
venn.compute_plot(named_sets, output_dir)
}
\arguments{
\item{named_sets}{A named list of character elements to include in Venn
diagram.}

\item{output_dir}{A string represent output directory to write files.}
}
\description{
This function ever write in files the intersection elements using convention
of \code{namedlist1_namedlist2_namedlist3.txt} and png image with the same
convention for full list of sets,
\code{namedlist1_namedlist2_namedlist3.png}.
}
\examples{
\dontrun{
a <- list(c("1", "2", "3"), c("1", "5"), c("1", "5", "7"))
names(a) <- c("a", "b", "c")
venn.compute_plot(a, "/home/user/Pictures")
b <- read.lists_from_files(c('genes_1.txt', 'genes_2.txt', 'genes_3.txt'), c("g1", "g2", "g3"))
venn.compute_plot(b, "pictures")
}
}
